experiments/requirements.txt
# V5.0 实验框架依赖

numpy>=1.20.0
pandas>=1.2.0
pyyaml>=5.4.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
pyarrow>=5.0.0  # 用于parquet格式
